ili9340: fix compile errors

+++ ili9340/rgb565.go
@@ -4,6 +4,5 @@ import "image/color"
 
 func rgb565(c color.Color) uint16 {
 	r, g, b, _ := c.RGBA()
-	// return ((r>>3)&0x1F)<<11 | ((g>>2)&0x3f)<<5 | (b>>3)&0x1F
-	return ((r & 0xF8) << 8) | ((g & 0xFC) << 3) | (b >> 3)
+	return (uint16(r&0xF8) << 8) | uint16((g&0xFC)<<3) | uint16(b>>3)
 }
